Memento pattern enhancement
